Revert "Merge branch 'update-password' of https://github.com/firebase/firebase-ios-sdk into add-linkedin"
This reverts commit c5ff4dda98f698c486c257d76262328a874190fc, reversing
changes made to 8111a7956e21bc1a0d25ea9fd604a26af648830f.

+++ Firestore/core/src/local/index_backfiller.cc
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include <algorithm>
-#include <string>
 #include <unordered_set>
 #include <utility>
 
@@ -45,7 +44,7 @@ IndexBackfiller::IndexBackfiller() {
   max_documents_to_process_ = kMaxDocumentsToProcess;
 }
 
-size_t IndexBackfiller::WriteIndexEntries(const LocalStore* local_store) {
+int IndexBackfiller::WriteIndexEntries(const LocalStore* local_store) {
   IndexManager* index_manager = local_store->index_manager();
   std::unordered_set<std::string> processed_collection_groups;
   size_t documents_remaining = max_documents_to_process_;
@@ -65,10 +64,10 @@ size_t IndexBackfiller::WriteIndexEntries(const LocalStore* local_store) {
   return max_documents_to_process_ - documents_remaining;
 }
 
-size_t IndexBackfiller::WriteEntriesForCollectionGroup(
+int IndexBackfiller::WriteEntriesForCollectionGroup(
     const LocalStore* local_store,
     const std::string& collection_group,
-    size_t documents_remaining_under_cap) const {
+    int documents_remaining_under_cap) const {
   IndexManager* index_manager = local_store->index_manager();
   const auto* const local_documents_view = local_store->local_documents();
